experiment/command-ms/requirements.txt
flask==2.1.1               # For creating the Flask web application
kafka-python==2.0.2        # For interacting with Kafka
